Split acs_device_info into multiple handlers
