+++ controllers/acs_device_info.js
@@ -2,47 +2,30 @@
 /* global __line */
 const DevicesAPI = require('./external-genieacs/devices-api');
 const TasksAPI = require('./external-genieacs/tasks-api');
-const FirmwaresAPI = require('./external-genieacs/firmwares-api');
 const controlApi = require('./external-api/control');
 const DeviceModel = require('../models/device');
 const DeviceVersion = require('../models/device_version');
-const FirmwareModel = require('../models/firmware');
 const Notification = require('../models/notification');
 const Config = require('../models/config');
 const sio = require('../sio');
 const deviceHandlers = require('./handlers/devices');
 const meshHandlers = require('./handlers/mesh');
-const acsHandlers = require('./handlers/acs');
+const acsFirmwareHandler = require('./handlers/acs/firmware.js')
+const acsAccessControlHandler = require('./handlers/acs/access_control.js');
+const acsPortForwardHandler = require('./handlers/acs/port_forward.js');
+const acsDiagnosticsHandler = require('./handlers/acs/diagnostics.js');
+const acsMeshDeviceHandler = require('./handlers/acs/mesh.js')
+const acsDeviceLogsHandler = require('./handlers/acs/logs.js');
+const acsConnDevicesHandler = require('./handlers/acs/connected_devices.js');
+const acsMeasuresHandler = require('./handlers/acs/measures.js');
+const acsXMLConfigHandler = require('./handlers/acs/xmlconfig.js');
 const utilHandlers = require('./handlers/util');
 const debug = require('debug')('ACS_DEVICE_INFO');
 const t = require('./language').i18next.t;
 
-const pako = require('pako');
 const http = require('http');
 
 let acsDeviceInfoController = {};
-
-const checkForNestedKey = function(data, key) {
-  if (!data) return false;
-  let current = data;
-  let splitKey = key.split('.');
-  for (let i = 0; i < splitKey.length; i++) {
-    if (!current.hasOwnProperty(splitKey[i])) return false;
-    current = current[splitKey[i]];
-  }
-  return true;
-};
-
-const getFromNestedKey = function(data, key) {
-  if (!data) return undefined;
-  let current = data;
-  let splitKey = key.split('.');
-  for (let i = 0; i < splitKey.length; i++) {
-    if (!current.hasOwnProperty(splitKey[i])) return undefined;
-    current = current[splitKey[i]];
-  }
-  return current;
-};
 
 const convertSubnetMaskToInt = function(mask) {
   if (mask === '255.255.255.0') {
@@ -104,36 +87,6 @@ const convertWifiMode = function(mode, is5ghz) {
     case 'ax':
     default:
       return undefined;
-  }
-};
-
-const convertToDbm = function(model, rxPower) {
-  switch (model) {
-    case 'HG9': {
-      rxPower = parseFloat(rxPower.split(' ')[0]);
-      if (isNaN(rxPower)) {
-        debug('rxPower is not an number!!!');
-      }
-      return rxPower;
-    }
-    case 'IGD':
-    case 'FW323DAC':
-    case 'F660':
-    case 'F670L':
-    case 'F680':
-    case 'ST-1001-FL':
-    case 'G-140W-C':
-    case 'G-140W-CS':
-    case 'G-140W-UD': {
-      rxPower = parseFloat((10 * Math.log10(rxPower*0.0001)).toFixed(3));
-      if (isNaN(rxPower)) {
-        debug('rxPower is not a number!!!');
-      }
-      return rxPower;
-    }
-    case 'GONUAC001':
-    default:
-      return rxPower;
   }
 };
 
@@ -204,58 +157,6 @@ const extractGreatekCredentials = function(config) {
   return {username: username, password: password};
 };
 
-const appendBytesMeasure = function(original, recv, sent) {
-  if (!original) original = {};
-  try {
-    let now = Math.floor(Date.now()/1000);
-    let bytes = JSON.parse(JSON.stringify(original));
-    if (Object.keys(bytes).length >= 300) {
-      let keysNum = Object
-        .keys(bytes)
-        .map((keyNum) => {
-          const parsedKeyNum = parseInt(keyNum);
-          if (isNaN(parsedKeyNum)) {
-            debug('parsedKeyNum is NaN!!!');
-          }
-          return parsedKeyNum;
-        });
-      let smallest = Math.min(...keysNum);
-      delete bytes[smallest];
-    }
-    bytes[now] = [recv, sent];
-    return bytes;
-  } catch (e) {
-    debug(`appendBytesMeasure Exception: ${e}`);
-    return original;
-  }
-};
-
-const appendPonSignal = function(original, rxPower, txPower) {
-  if (!original) original = {};
-  try {
-    let now = Math.floor(Date.now() / 1000);
-    let dbms = JSON.parse(JSON.stringify(original));
-    if (Object.keys(dbms).length >= 100) {
-      let keysNum = Object
-        .keys(dbms)
-        .map((keyNum) => {
-          const parsedKeyNum = parseInt(keyNum);
-          if (isNaN(parsedKeyNum)) {
-            debug('parsedKeyNum is NaN!!!');
-          }
-          return parsedKeyNum;
-        });
-      let smallest = Math.min(...keysNum);
-      delete dbms[smallest];
-    }
-    dbms[now] = [rxPower, txPower];
-    return dbms;
-  } catch (e) {
-    debug(`appendPonSignal Exception: ${e}`);
-    return original;
-  }
-};
-
 const processHostFromURL = function(url) {
   if (typeof url !== 'string') return '';
   let doubleSlash = url.indexOf('//');
@@ -270,45 +171,6 @@ const processHostFromURL = function(url) {
   let endIndex = (pathStart >= 0) ? doubleSlash+2+pathStart : url.length;
   let hostAndPort = url.substring(doubleSlash+2, endIndex);
   return hostAndPort.split(':')[0];
-};
-
-const saveDeviceData = async function(mac, landevices) {
-  if (!mac || !landevices) return;
-  let device = await DeviceModel.findById(mac.toUpperCase());
-  if (!device) return;
-  landevices.forEach((lanDev)=>{
-    let lanMac = lanDev.mac.toUpperCase();
-    let registered = device.lan_devices.find((d)=>d.mac===lanMac);
-    if (registered) {
-      registered.dhcp_name = lanDev.name;
-      registered.ip = lanDev.ip;
-      registered.conn_type = (lanDev.wifi) ? 1 : 0;
-      if (lanDev.rate) registered.conn_speed = lanDev.rate;
-      if (lanDev.wifi_freq) registered.wifi_freq = lanDev.wifi_freq;
-      if (lanDev.rssi) registered.wifi_signal = lanDev.rssi;
-      if (lanDev.snr) registered.wifi_snr = lanDev.snr;
-      if (lanDev.wifi_mode) registered.wifi_mode = lanDev.wifi_mode;
-      registered.last_seen = Date.now();
-    } else {
-      device.lan_devices.push({
-        mac: lanMac,
-        dhcp_name: lanDev.name,
-        ip: lanDev.ip,
-        conn_type: (lanDev.wifi) ? 1 : 0,
-        conn_speed: (lanDev.rate) ? lanDev.rate : undefined,
-        wifi_signal: (lanDev.rssi) ? lanDev.rssi : undefined,
-        wifi_freq: (lanDev.wifi_freq) ? lanDev.wifi_freq : undefined,
-        wifi_snr: (lanDev.snr) ? lanDev.snr : undefined,
-        wifi_mode: (lanDev.wifi_mode) ? lanDev.wifi_mode : undefined,
-        last_seen: Date.now(),
-        first_seen: Date.now(),
-      });
-    }
-  });
-  device.last_devices_refresh = Date.now();
-  await device.save().catch((err) => {
-    console.log('Error saving tr-069 device data ' + mac + ': ' + err);
-  });
 };
 
 const createRegistry = async function(req, permissions) {
@@ -987,7 +849,7 @@ acsDeviceInfoController.syncDevice = async function(req, res) {
   }
   if (data.wan.recv_bytes && data.wan.recv_bytes.value &&
       data.wan.sent_bytes && data.wan.sent_bytes.value) {
-    device.wan_bytes = appendBytesMeasure(
+    device.wan_bytes = acsMeasuresHandler.appendBytesMeasure(
       device.wan_bytes,
       data.wan.recv_bytes.value,
       data.wan.sent_bytes.value,
@@ -996,25 +858,29 @@ acsDeviceInfoController.syncDevice = async function(req, res) {
   let isPonRxValOk = false;
   let isPonTxValOk = false;
   if (data.wan.pon_rxpower && data.wan.pon_rxpower.value) {
-    device.pon_rxpower = convertToDbm(data.common.model.value,
-                                      data.wan.pon_rxpower.value);
+    device.pon_rxpower = acsMeasuresHandler.convertToDbm(
+      data.common.model.value, data.wan.pon_rxpower.value,
+    );
     isPonRxValOk = true;
   } else if (data.wan.pon_rxpower_epon && data.wan.pon_rxpower_epon.value) {
-    device.pon_rxpower = convertToDbm(data.common.model.value,
-                                      data.wan.pon_rxpower_epon.value);
+    device.pon_rxpower = acsMeasuresHandler.convertToDbm(
+      data.common.model.value, data.wan.pon_rxpower_epon.value,
+    );
     isPonRxValOk = true;
   }
   if (data.wan.pon_txpower && data.wan.pon_txpower.value) {
-    device.pon_txpower = convertToDbm(data.common.model.value,
-                                      data.wan.pon_txpower.value);
+    device.pon_txpower = acsMeasuresHandler.convertToDbm(
+      data.common.model.value, data.wan.pon_txpower.value,
+    );
     isPonTxValOk = true;
   } else if (data.wan.pon_txpower_epon && data.wan.pon_txpower_epon.value) {
-    device.pon_txpower = convertToDbm(data.common.model.value,
-                                      data.wan.pon_txpower_epon.value);
+    device.pon_txpower = acsMeasuresHandler.convertToDbm(
+      data.common.model.value, data.wan.pon_txpower_epon.value,
+    );
     isPonTxValOk = true;
   }
   if (isPonRxValOk && isPonTxValOk) {
-    device.pon_signal_measure = appendPonSignal(
+    device.pon_signal_measure = acsMeasuresHandler.appendPonSignal(
       device.pon_signal_measure,
       device.pon_rxpower,
       device.pon_txpower,
@@ -1140,9 +1006,9 @@ acsDeviceInfoController.syncDevice = async function(req, res) {
         if (entriesDiff != 0) {
           // If entries sizes are not the same, no need to check
           // entry by entry differences
-          acsDeviceInfoController.changePortForwardRules(device, entriesDiff);
+          acsPortForwardHandler.changePortForwardRules(device, entriesDiff);
         } else {
-          acsDeviceInfoController.checkPortForwardRules(device);
+          acsPortForwardHandler.checkPortForwardRules(device);
         }
       }
     }
@@ -1163,7 +1029,7 @@ acsDeviceInfoController.syncDevice = async function(req, res) {
         device.web_admin_user = 'root';
       }
       targets.push('web-admin');
-      configFileEditing(device, targets);
+      acsXMLConfigHandler.configFileEditing(device, targets);
     } else {
       // Send web admin password correct setup for those CPEs that always
       // retrieve blank on this field
@@ -1209,63 +1075,6 @@ acsDeviceInfoController.rebootDevice = async function(device, res) {
   }
 };
 
-// TODO: Move this function to external-genieacs?
-const fetchLogFromGenie = async function(acsID) {
-  let device;
-  try {
-    device = await DeviceModel.findOne({acs_id: acsID}).lean();
-  } catch (e) {
-    return;
-  }
-  if (!device || !device.use_tr069) {
-    return;
-  }
-  let mac = device._id;
-  let logField = DevicesAPI.getModelFieldsFromDevice(device).fields.log;
-  let query = {_id: acsID};
-  let path = '/devices/?query='+JSON.stringify(query)+'&projection='+logField;
-  let options = {
-    method: 'GET',
-    hostname: 'localhost',
-    port: 7557,
-    path: encodeURI(path),
-  };
-  let req = http.request(options, (resp)=>{
-    resp.setEncoding('utf8');
-    let data = '';
-    resp.on('data', (chunk)=>data+=chunk);
-    resp.on('end', async () => {
-      if (data.length > 0) {
-        try {
-          data = JSON.parse(data)[0];
-        } catch (err) {
-          debug(err);
-          data = '';
-        }
-      }
-      let success = false;
-      if (!checkForNestedKey(data, logField+'._value')) {
-        data = t('logUnavailable', {errorline: __line});
-      } else {
-        success = true;
-        data = getFromNestedKey(data, logField+'._value');
-      }
-      let compressedLog = pako.gzip(data);
-      if (success) {
-        let deviceEdit = await DeviceModel.findById(mac);
-        deviceEdit.last_contact = Date.now();
-        deviceEdit.lastboot_date = Date.now();
-        deviceEdit.lastboot_log = Buffer.from(compressedLog);
-        await deviceEdit.save().catch((err) => {
-          console.log('Error saving last boot log to database: ' + err);
-        });
-      }
-      sio.anlixSendLiveLogNotifications(mac, compressedLog);
-    });
-  });
-  req.end();
-};
-
 acsDeviceInfoController.requestDiagnosticsResults = async function(req, res) {
   let acsID = req.body.acs_id;
   let device;
@@ -1290,1074 +1099,17 @@ acsDeviceInfoController.requestDiagnosticsResults = async function(req, res) {
       fields.diagnostics.ping.root, fields.diagnostics.speedtest.root,
     ],
   };
-  TasksAPI.addTask(acsID, task, fetchDiagnosticsFromGenie);
+  TasksAPI.addTask(
+    acsID, task, acsDiagnosticsHandler.fetchDiagnosticsFromGenie,
+  );
 }
 
-const fetchDiagnosticsFromGenie = async function(acsID) {
-  let device;
-  try {
-    device = await DeviceModel.findOne({acs_id: acsID}).lean();
-  } catch (e) {
-    return;
-  }
-  if (!device || !device.use_tr069 || !device.acs_id) {
-    return;
-  }
-
-  let parameters = [];
-  let diagNecessaryKeys = {
-    ping: {
-      diag_state: '',
-      num_of_rep: '',
-      failure_count: '',
-      success_count: '',
-      host: '',
-      avg_resp_time: '',
-      max_resp_time: '',
-      min_resp_time: '',
-    },
-    speedtest: {
-      diag_state: '',
-      num_of_conn: '',
-      download_url: '',
-      bgn_time: '',
-      end_time: '',
-      test_bytes_rec: '',
-      down_transports: '',
-      full_load_bytes_rec: '',
-      full_load_period: '',
-    },
-  };
-
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-
-  for (let masterKey in diagNecessaryKeys) {
-    if (
-      diagNecessaryKeys.hasOwnProperty(masterKey) &&
-      fields.diagnostics.hasOwnProperty(masterKey)
-    ) {
-      let keys = diagNecessaryKeys[masterKey];
-      let genieFields = fields.diagnostics[masterKey];
-      for (let key in keys) {
-        if (genieFields.hasOwnProperty(key)) {
-          parameters.push(genieFields[key]);
-        }
-      }
-    }
-  }
-
-  let success = false;
-  let query = {_id: acsID};
-  let path = '/devices/?query='+JSON.stringify(query)+
-              '&projection='+parameters.join(',');
-  let options = {
-    protocol: 'http:',
-    method: 'GET',
-    hostname: 'localhost',
-    port: 7557,
-    path: encodeURI(path),
-  };
-  let request = http.request(options, (response)=>{
-    let chunks = [];
-    response.on('error', (error) => console.log(error));
-    response.on('data', async (chunk)=>chunks.push(chunk));
-    response.on('end', async (chunk) => {
-      let body = Buffer.concat(chunks);
-      try {
-        let data = JSON.parse(body)[0];
-        let permissions = DeviceVersion.findByVersion(
-          device.version,
-          device.wifi_is_5ghz_capable,
-          device.model,
-        );
-        if (permissions) {
-          if (permissions.grantPingTest) {
-            await acsDeviceInfoController.calculatePingDiagnostic(
-              device, data, diagNecessaryKeys.ping, fields.diagnostics.ping,
-            );
-          }
-          if (permissions.grantSpeedTest) {
-            await acsDeviceInfoController.calculateSpeedDiagnostic(
-              device, data, diagNecessaryKeys.speedtest,
-              fields.diagnostics.speedtest,
-            );
-          }
-        } else {
-          console.log('Failed: genie can\'t check device permissions');
-        }
-      } catch (e) {
-        console.log('Failed: genie response was not valid');
-        console.log('Error:', e);
-      }
-    });
-  });
-  request.end();
-};
-
-acsDeviceInfoController.getAllNestedKeysFromObject = function(
-  data, target, genieFields,
-) {
-  let result = {};
-  Object.keys(target).forEach((key)=>{
-    if (checkForNestedKey(data, genieFields[key]+'._value')) {
-      result[key] = getFromNestedKey(data, genieFields[key]+'._value');
-    }
-  });
-  return result;
-};
-
 acsDeviceInfoController.firePingDiagnose = async function(mac) {
-  let device;
-  try {
-    device = await DeviceModel.findById(mac).lean();
-  } catch (e) {
-    console.log('Error:', e);
-    return {success: false,
-            message: t('cpeFindError', {errorline: __line})};
-  }
-  if (!device || !device.use_tr069 || !device.acs_id) {
-    return {success: false,
-            message: t('cpeFindError', {errorline: __line})};
-  }
-  let acsID = device.acs_id;
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  let diagnIPPingDiagnostics = fields.diagnostics.ping.root;
-  // We need to update the parameter values before we fire the ping test
-  let task = {
-    name: 'getParameterValues',
-    parameterNames: [diagnIPPingDiagnostics],
-  };
-  const result = await TasksAPI.addTask(acsID, task, startPingDiagnose);
-  if (result.success) {
-    return {success: true, message: t('operationSuccessful')};
-  } else {
-    return {
-      success: false, message: t('acsPingCouldNotBeSent', {errorline: __line}),
-    };
-  }
-};
-
-const startPingDiagnose = async function(acsID) {
-  let device;
-  try {
-    device = await DeviceModel.findOne({acs_id: acsID}).lean();
-  } catch (err) {
-    return {success: false, message: err.message + ' in ' + acsID};
-  }
-  if (!device) {
-    return {success: false, message: t('cpeFindError', {errorline: __line})};
-  }
-
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  let diagnStateField = fields.diagnostics.ping.diag_state;
-  let diagnNumRepField = fields.diagnostics.ping.num_of_rep;
-  let diagnURLField = fields.diagnostics.ping.host;
-  let diagnTimeoutField = fields.diagnostics.ping.timeout;
-
-  let numberOfRep = 10;
-  let pingHostUrl = device.ping_hosts[0];
-  let timeout = 1000;
-
-  let task = {
-    name: 'setParameterValues',
-    parameterValues: [[diagnStateField, 'Requested', 'xsd:string'],
-                      [diagnNumRepField, numberOfRep, 'xsd:unsignedInt'],
-                      [diagnURLField, pingHostUrl, 'xsd:string'],
-                      [diagnTimeoutField, timeout, 'xsd:unsignedInt']],
-  };
-  const result = await TasksAPI.addTask(acsID, task);
-  if (!result.success) {
-    console.log('Error starting ping diagnose for ' + acsID);
-  }
-};
-
-acsDeviceInfoController.calculatePingDiagnostic = function(device, data,
-                                                           pingKeys,
-                                                           pingFields) {
-  pingKeys = acsDeviceInfoController.getAllNestedKeysFromObject(
-    data, pingKeys, pingFields,
-  );
-  if (pingKeys.diag_state !== 'Requested' && pingKeys.diag_state !== 'None') {
-    let result = {};
-    device.ping_hosts.forEach((host) => {
-      if (host) {
-        result[host] = {
-          lat: '---',
-          loss: '--- ',
-        };
-      }
-    });
-    if (
-      pingKeys.diag_state === 'Complete' ||
-      pingKeys.diag_state === 'Complete\n'
-    ) {
-      const loss = parseInt(pingKeys.failure_count * 100 /
-        (pingKeys.success_count + pingKeys.failure_count));
-      if (isNaN(loss)) {
-        debug('calculatePingDiagnostic loss is not an number!!!');
-      }
-      result[pingKeys.host] = {
-        lat: pingKeys.avg_resp_time.toString(),
-        loss: loss.toString(),
-      };
-      let model = device.model;
-      if (model === 'HG8245Q2' || model === 'EG8145V5' || model === 'HG9') {
-        if (pingKeys.success_count === 1) result[pingKeys.host]['loss'] = '0';
-        else result[pingKeys.host]['loss'] = '100';
-      }
-    }
-    deviceHandlers.sendPingToTraps(device._id, {results: result});
-  }
-};
-
-acsDeviceInfoController.getSpeedtestFile = async function(device) {
-  let matchedConfig = await Config.findOne(
-    {is_default: true}, {measureServerIP: true, measureServerPort: true},
-  ).lean().catch(
-    function(err) {
-      console.error('Error creating entry: ' + err);
-      return '';
-    },
-  );
-  if (!matchedConfig) {
-    console.error('Error creating entry. Config does not exists.');
-    return '';
-  }
-  let stage = device.current_speedtest.stage;
-  let band = device.current_speedtest.band_estimative;
-  let url = 'http://' + matchedConfig.measureServerIP + ':' +
-                  matchedConfig.measureServerPort + '/measure/tr069/';
-  if (stage) {
-    if (stage == 'estimative') {
-      return url + 'file_512KB.bin';
-    }
-    if (stage == 'measure') {
-      if (band >= 700) {
-        return url + 'file_640000KB.bin';
-      } else if (band >= 500) {
-        return url + 'file_448000KB.bin';
-      } else if (band >= 300) {
-        return url + 'file_320000KB.bin';
-      } else if (band >= 100) {
-        return url + 'file_192000KB.bin';
-      } else if (band >= 50) {
-        return url + 'file_64000KB.bin';
-      } else if (band >= 30) {
-        return url + 'file_32000KB.bin';
-      } else if (band >= 10) {
-        return url + 'file_19200KB.bin';
-      } else if (band >= 5) {
-        return url + 'file_6400KB.bin';
-      } else if (band >= 3) {
-        return url + 'file_1920KB.bin';
-      } else if (band < 3) {
-        return url + 'file_512KB.bin';
-      }
-    }
-  }
-  return '';
+  return await acsDiagnosticsHandler.firePingDiagnose(mac);
 };
 
 acsDeviceInfoController.fireSpeedDiagnose = async function(mac) {
-  let device;
-  try {
-    device = await DeviceModel.findById(mac).lean();
-  } catch (e) {
-    console.log('Error:', e);
-    return {success: false,
-            message: t('cpeFindError', {errorline: __line})};
-  }
-  if (!device || !device.use_tr069 || !device.acs_id) {
-    return {success: false,
-            message: t('cpeFindError', {errorline: __line})};
-  }
-  let acsID = device.acs_id;
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  let diagnSpeedtestDiagnostics = fields.diagnostics.speedtest.root;
-  // We need to update the parameter values before we fire the speedtest
-  let task = {
-    name: 'getParameterValues',
-    parameterNames: [diagnSpeedtestDiagnostics],
-  };
-  const result = await TasksAPI.addTask(acsID, task, startSpeedtestDiagnose);
-  if (result.success) {
-    return {success: true, message: t('operationSuccessful')};
-  } else {
-    return {
-      success: false, message: t('acsSpeedTestError', {errorline: __line}),
-    };
-  }
-};
-
-const startSpeedtestDiagnose = async function(acsID) {
-  let device;
-  try {
-    device = await DeviceModel.findOne({acs_id: acsID}).lean();
-  } catch (err) {
-    return {success: false, message: err.message + ' in ' + acsID};
-  }
-  if (!device) {
-    return {success: false, message: t('cpeFindError', {errorline: __line})};
-  }
-
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  let diagnStateField = fields.diagnostics.speedtest.diag_state;
-  let diagnNumConnField = fields.diagnostics.speedtest.num_of_conn;
-  let diagnURLField = fields.diagnostics.speedtest.download_url;
-
-  let numberOfCon = 3;
-  let speedtestHostUrl = await acsDeviceInfoController.getSpeedtestFile(device);
-
-  if (!speedtestHostUrl || speedtestHostUrl === '') {
-    console.log('No valid speedtest URL found for ' + acsID);
-    return;
-  }
-
-  let task = {
-    name: 'setParameterValues',
-    parameterValues: [[diagnStateField, 'Requested', 'xsd:string'],
-                      [diagnNumConnField, numberOfCon, 'xsd:unsignedInt'],
-                      [diagnURLField, speedtestHostUrl, 'xsd:string']],
-  };
-  const result = await TasksAPI.addTask(acsID, task);
-  if (!result.success) {
-    console.log('Error starting speedtest diagnose for ' + acsID);
-  }
-};
-
-acsDeviceInfoController.calculateSpeedDiagnostic = async function(device, data,
-                                                                  speedKeys,
-                                                                  speedFields) {
-  speedKeys = acsDeviceInfoController.getAllNestedKeysFromObject(
-    data, speedKeys, speedFields,
-  );
-  let result;
-  let speedValueBasic;
-  let speedValueFullLoad;
-  let rqstTime;
-  let lastTime = (new Date(1970, 0, 1)).valueOf();
-
-  try {
-    if ('current_speedtest' in device &&
-        'timestamp' in device.current_speedtest &&
-        device.current_speedtest.timestamp) {
-      rqstTime = device.current_speedtest.timestamp.valueOf();
-    }
-  } catch (e) {
-    console.log('Error at TR-069 speedtest:', e);
-    return;
-  }
-
-  if (!device.current_speedtest.timestamp || (rqstTime > lastTime)) {
-    if (speedKeys.diag_state == 'Completed') {
-      if (device.speedtest_results.length > 0) {
-        lastTime = utilHandlers.parseDate(
-          device.speedtest_results[device.speedtest_results.length-1].timestamp,
-        );
-      }
-
-      let beginTime = (new Date(speedKeys.bgn_time)).valueOf();
-      let endTime = (new Date(speedKeys.end_time)).valueOf();
-      // 10**3 => seconds to miliseconds (because of valueOf() notation)
-      let deltaTime = (endTime - beginTime) / (10**3);
-
-      // 8 => byte to bit
-      // 1024**2 => bit to megabit
-      speedValueBasic = (8/(1024**2))*(speedKeys.test_bytes_rec/deltaTime);
-
-      if (speedKeys.full_load_bytes_rec && speedKeys.full_load_period) {
-        // 10**6 => microsecond to second
-        // 8 => byte to bit
-        // 1024**2 => bit to megabit
-        speedValueFullLoad = ((8*(10**6))/(1024**2)) *
-                    (speedKeys.full_load_bytes_rec/speedKeys.full_load_period);
-      }
-
-      // Speedtest's estimative / real measure step
-      if (device.current_speedtest.stage == 'estimative') {
-        device.current_speedtest.band_estimative = speedValueBasic;
-        device.current_speedtest.stage = 'measure';
-        await device.save().catch((err) => {
-          console.log('Error saving speed test est to database: ' + err);
-        });
-        await sio.anlixSendSpeedTestNotifications(device._id, {
-          stage: 'estimative_finished',
-          user: device.current_speedtest.user,
-        });
-        acsDeviceInfoController.fireSpeedDiagnose(device._id);
-        return;
-      } else if (device.current_speedtest.stage == 'measure') {
-        result = {
-          downSpeed: '',
-          user: device.current_speedtest.user,
-        };
-        if (speedKeys.full_load_bytes_rec && speedKeys.full_load_period) {
-          result.downSpeed = parseInt(speedValueFullLoad).toString() + ' Mbps';
-          if (isNaN(parseInt(speedValueFullLoad))) {
-            result.downSpeed = '0 Mbps';
-          }
-        } else {
-          result.downSpeed = parseInt(speedValueBasic).toString() + ' Mbps';
-          if (isNaN(parseInt(speedValueBasic))) {
-            result.downSpeed = '0 Mbps';
-          }
-        }
-        deviceHandlers.storeSpeedtestResult(device, result);
-        return;
-      }
-    } else {
-      // Error treatment (switch-case for future error handling)
-      switch (speedKeys.diag_state) {
-        case 'Error_InitConnectionFailed':
-        case 'Error_NoResponse':
-        case 'Error_Other':
-          console.log('Failure at TR-069 speedtest:', speedKeys.diag_state);
-          result = {
-            downSpeed: '503 Server',
-            user: device.current_speedtest.user,
-          };
-          break;
-        default:
-          result = {
-            user: device.current_speedtest.user,
-          };
-      }
-      deviceHandlers.storeSpeedtestResult(device, result);
-      return;
-    }
-  }
-};
-
-// TODO: Move this function to external-genieacs?
-const fetchWanBytesFromGenie = async function(acsID) {
-  let device;
-  try {
-    device = await DeviceModel.findOne({acs_id: acsID}).lean();
-  } catch (e) {
-    return;
-  }
-  if (!device || !device.use_tr069) {
-    return;
-  }
-  let mac = device._id;
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  let recvField = fields.wan.recv_bytes;
-  let sentField = fields.wan.sent_bytes;
-  let query = {_id: acsID};
-  let projection = recvField + ',' + sentField;
-  let path = '/devices/?query='+JSON.stringify(query)+'&projection='+projection;
-  let options = {
-    method: 'GET',
-    hostname: 'localhost',
-    port: 7557,
-    path: encodeURI(path),
-  };
-  let req = http.request(options, (resp)=>{
-    resp.setEncoding('utf8');
-    let data = '';
-    let wanBytes = {};
-    resp.on('data', (chunk)=>data+=chunk);
-    resp.on('end', async () => {
-      if (data.length > 0) {
-        try {
-          data = JSON.parse(data)[0];
-        } catch (err) {
-          debug(err);
-          data = '';
-        }
-      }
-      let success = false;
-      if (checkForNestedKey(data, recvField+'._value') &&
-          checkForNestedKey(data, sentField+'._value')) {
-        success = true;
-        wanBytes = {
-          recv: getFromNestedKey(data, recvField+'._value'),
-          sent: getFromNestedKey(data, sentField+'._value'),
-        };
-      }
-      if (success) {
-        let deviceEdit = await DeviceModel.findById(mac);
-        if (!deviceEdit) return;
-        deviceEdit.last_contact = Date.now();
-        wanBytes = appendBytesMeasure(
-          deviceEdit.wan_bytes,
-          wanBytes.recv,
-          wanBytes.sent,
-        );
-        deviceEdit.wan_bytes = wanBytes;
-        await deviceEdit.save().catch((err) => {
-          console.log('Error saving device wan bytes: ' + err);
-        });
-      }
-      sio.anlixSendWanBytesNotification(mac, {wanbytes: wanBytes});
-    });
-  });
-  req.end();
-};
-
-const fetchUpStatusFromGenie = async function(acsID) {
-  let device;
-  try {
-    device = await DeviceModel.findOne({acs_id: acsID}).lean();
-  } catch (e) {
-    return;
-  }
-  if (!device || !device.use_tr069) {
-    return;
-  }
-  let mac = device._id;
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  let upTimeField1 = fields.wan.uptime.replace('*', 1);
-  let upTimeField2 = fields.wan.uptime.replace('*', 2);
-  let upTimePPPField1 = fields.wan.uptime_ppp.replace('*', 1).replace('*', 1);
-  let upTimePPPField2 = fields.wan.uptime_ppp.replace('*', 1).replace('*', 2);
-  let PPPoEUser1 = fields.wan.pppoe_user.replace('*', 1).replace('*', 1);
-  let PPPoEUser2 = fields.wan.pppoe_user.replace('*', 1).replace('*', 2);
-  let rxPowerField;
-  let txPowerField;
-  let rxPowerFieldEpon;
-  let txPowerFieldEpon;
-  let query = {_id: acsID};
-  let projection = fields.common.uptime +
-      ',' + upTimeField1 +
-      ',' + upTimeField2 +
-      ',' + upTimePPPField1 +
-      ',' + upTimePPPField2 +
-      ',' + PPPoEUser1 +
-      ',' + PPPoEUser2;
-
-  if (fields.wan.pon_rxpower && fields.wan.pon_txpower) {
-    rxPowerField = fields.wan.pon_rxpower;
-    txPowerField = fields.wan.pon_txpower;
-    projection += ',' + rxPowerField + ',' + txPowerField;
-  }
-
-  if (fields.wan.pon_rxpower_epon && fields.wan.pon_txpower_epon) {
-    rxPowerFieldEpon = fields.wan.pon_rxpower_epon;
-    txPowerFieldEpon = fields.wan.pon_txpower_epon;
-    projection += ',' + rxPowerFieldEpon + ',' + txPowerFieldEpon;
-  }
-  let path = '/devices/?query='+JSON.stringify(query)+'&projection='+projection;
-  let options = {
-    method: 'GET',
-    hostname: 'localhost',
-    port: 7557,
-    path: encodeURI(path),
-  };
-  let req = http.request(options, (resp)=>{
-    resp.setEncoding('utf8');
-    let data = '';
-    let sysUpTime = 0;
-    let wanUpTime = 0;
-    let signalState = {};
-    let ponSignal = {};
-    resp.on('data', (chunk)=>data+=chunk);
-    resp.on('end', async () => {
-      if (data.length > 0) {
-        try {
-          data = JSON.parse(data)[0];
-        } catch (err) {
-          debug(err);
-          return;
-        }
-      }
-      let successSys = false;
-      let successWan = false;
-      let successRxPower = false;
-      if (checkForNestedKey(data, fields.common.uptime+'._value')) {
-        successSys = true;
-        sysUpTime = getFromNestedKey(data, fields.common.uptime+'._value');
-      }
-      if (checkForNestedKey(data, PPPoEUser1+'._value')) {
-        successWan = true;
-        let hasPPPoE = getFromNestedKey(data, PPPoEUser1+'._value');
-        if (hasPPPoE && checkForNestedKey(data, upTimePPPField1+'._value')) {
-          wanUpTime = getFromNestedKey(data, upTimePPPField1+'._value');
-        }
-      } else if (checkForNestedKey(data, PPPoEUser2+'._value')) {
-        successWan = true;
-        let hasPPPoE = getFromNestedKey(data, PPPoEUser2+'._value');
-        if (hasPPPoE && checkForNestedKey(data, upTimePPPField2+'._value')) {
-          wanUpTime = getFromNestedKey(data, upTimePPPField2+'._value');
-        }
-      } else if (checkForNestedKey(data, upTimeField1+'._value')) {
-        successWan = true;
-        wanUpTime = getFromNestedKey(data, upTimeField1+'._value');
-      } else if (checkForNestedKey(data, upTimeField2+'._value')) {
-        successWan = true;
-        wanUpTime = getFromNestedKey(data, upTimeField2+'._value');
-      }
-      if (checkForNestedKey(data, rxPowerField + '._value') &&
-          checkForNestedKey(data, txPowerField + '._value')) {
-        successRxPower = true;
-        ponSignal = {
-          rxpower: getFromNestedKey(data, rxPowerField + '._value'),
-          txpower: getFromNestedKey(data, txPowerField + '._value'),
-        };
-      } else if (checkForNestedKey(data, rxPowerFieldEpon + '._value') &&
-                 checkForNestedKey(data, txPowerFieldEpon + '._value')) {
-        successRxPower = true;
-        ponSignal = {
-          rxpower: getFromNestedKey(data, rxPowerFieldEpon + '._value'),
-          txpower: getFromNestedKey(data, txPowerFieldEpon + '._value'),
-        };
-      }
-      if (successSys || successWan || successRxPower) {
-        let deviceEdit = await DeviceModel.findById(mac);
-        deviceEdit.last_contact = Date.now();
-        deviceEdit.sys_up_time = sysUpTime;
-        deviceEdit.wan_up_time = wanUpTime;
-        if (successRxPower) {
-          // covert rx and tx signal
-          ponSignal.rxpower = convertToDbm(deviceEdit.model, ponSignal.rxpower);
-          ponSignal.txpower = convertToDbm(deviceEdit.model, ponSignal.txpower);
-          // send then
-          let config = await Config.findOne(
-            {is_default: true}, {tr069: true},
-          ).lean();
-          signalState = {
-            rxpower: ponSignal.rxpower,
-            threshold:
-              config.tr069.pon_signal_threshold,
-            thresholdCritical:
-              config.tr069.pon_signal_threshold_critical,
-            thresholdCriticalHigh:
-              config.tr069.pon_signal_threshold_critical_high,
-          };
-          // append to device data structure
-          ponSignal = appendPonSignal(
-            deviceEdit.pon_signal_measure,
-            ponSignal.rxpower,
-            ponSignal.txpower,
-          );
-          deviceEdit.pon_signal_measure = ponSignal;
-        }
-        await deviceEdit.save().catch((err) => {
-          console.log('Error saving device up status: ' + err);
-        });
-      }
-      sio.anlixSendUpStatusNotification(mac, {
-        sysuptime: sysUpTime,
-        wanuptime: wanUpTime,
-        ponsignal: signalState,
-      });
-    });
-  });
-  req.end();
-};
-
-const checkMeshObjsCreated = function(device) {
-  return new Promise((resolve, reject) => {
-    let acsID = device.acs_id;
-    let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-    let query = {_id: acsID};
-    let projection = `${fields.mesh2.ssid}, ${fields.mesh5.ssid}`;
-    let path =
-      `/devices/?query=${JSON.stringify(query)}&projection=${projection}`;
-    let options = {
-      method: 'GET',
-      hostname: 'localhost',
-      port: 7557,
-      path: encodeURI(path),
-    };
-    let result = {
-      mesh2: false,
-      mesh5: false,
-      success: true,
-    };
-    let req = http.request(options, (resp)=>{
-      resp.setEncoding('utf8');
-      let data = '';
-      resp.on('data', (chunk)=>data+=chunk);
-      resp.on('end', async () => {
-        try {
-          data = JSON.parse(data)[0];
-        } catch (e) {
-          result.success = false;
-          resolve(result);
-        }
-        if (checkForNestedKey(data, `${fields.mesh2.ssid}._value`)) {
-          result.mesh2 = true;
-        }
-        if (checkForNestedKey(data, `${fields.mesh5.ssid}._value`)) {
-          result.mesh5 = true;
-        }
-        resolve(result);
-      });
-    });
-    req.end();
-  });
-};
-
-const fetchMeshBSSID = function(device, meshMode) {
-  return new Promise((resolve, reject) => {
-    let acsID = device.acs_id;
-    let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-    let query = {_id: acsID};
-    let projection = '';
-    if (meshMode === 2 || meshMode === 4) {
-      projection += `${fields.mesh2.bssid},`;
-    }
-    if (meshMode === 3 || meshMode === 4) {
-      projection += `${fields.mesh5.bssid},`;
-    }
-    // Removing trailing comma from projection
-    projection = projection.slice(0, -1);
-    let path =
-      `/devices/?query=${JSON.stringify(query)}&projection=${projection}`;
-    let options = {
-      method: 'GET',
-      hostname: 'localhost',
-      port: 7557,
-      path: encodeURI(path),
-    };
-    let req = http.request(options, (resp)=>{
-      resp.setEncoding('utf8');
-      let data = '';
-      resp.on('data', (chunk)=>data+=chunk);
-      resp.on('end', async () => {
-        try {
-          data = JSON.parse(data)[0];
-        } catch (e) {
-          console.log('Error parsing bssid data from genie');
-          return resolve({success: false});
-        }
-        let bssid2 = '';
-        let bssid5 = '';
-        // Mesh modes that use 2.4GHz radio
-        if (meshMode === 2 || meshMode === 4) {
-          // Check if field exists and collect it from genie
-          let field = `${fields.mesh2.bssid}._value`;
-          if (checkForNestedKey(data, field)) {
-            bssid2 = getFromNestedKey(data, field);
-          }
-          // We need to make sure bssid2 is not empty and different than 0
-          if (!bssid2 || bssid2 === '00:00:00:00:00:00') {
-            return {success: false};
-          }
-        }
-        // Mesh modes that use 5Hz radio
-        if (meshMode === 3 || meshMode === 4) {
-          // Check if field exists and collect it from genie
-          let field = `${fields.mesh5.bssid}._value`;
-          if (checkForNestedKey(data, field)) {
-            bssid5 = getFromNestedKey(data, field);
-          }
-          // We need to make sure bssid5 is not empty and different than 0
-          if (!bssid5 || bssid5 === '00:00:00:00:00:00') {
-            return {success: false};
-          }
-        }
-        resolve({sucess: true, mesh2: bssid2, mesh5: bssid5});
-      });
-    });
-    req.end();
-  });
-};
-
-// TODO: Move this function to external-genieacs?
-acsDeviceInfoController.fetchPonSignalFromGenie = async function(acsID) {
-  let device;
-  try {
-    device = await DeviceModel.findOne({acs_id: acsID}).lean();
-  } catch (e) {
-    console.log('Error:', e);
-    return {success: false,
-            message: t('cpeFindError', {errorline: __line})};
-  }
-  let mac = device._id;
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  let rxPowerField = fields.wan.pon_rxpower;
-  let txPowerField = fields.wan.pon_txpower;
-  let rxPowerFieldEpon = '';
-  let txPowerFieldEpon = '';
-  let projection = rxPowerField + ',' + txPowerField;
-
-  if (fields.wan.pon_rxpower_epon && fields.wan.pon_txpower_epon) {
-    rxPowerFieldEpon = fields.wan.pon_rxpower_epon;
-    txPowerFieldEpon = fields.wan.pon_txpower_epon;
-    projection += ',' + rxPowerFieldEpon + ',' + txPowerFieldEpon;
-  }
-
-  let query = {_id: acsID};
-  let path = '/devices/?query='+JSON.stringify(query)+'&projection='+projection;
-  let options = {
-    method: 'GET',
-    hostname: 'localhost',
-    port: 7557,
-    path: encodeURI(path),
-  };
-  let req = http.request(options, (resp)=>{
-    resp.setEncoding('utf8');
-    let data = '';
-    let ponSignal = {};
-    resp.on('data', (chunk)=>data+=chunk);
-    resp.on('end', async () => {
-      if (data.length > 0) {
-        try {
-          data = JSON.parse(data)[0];
-        } catch (err) {
-          debug(err);
-          data = '';
-        }
-      }
-      let success = false;
-      if (checkForNestedKey(data, rxPowerField + '._value') &&
-          checkForNestedKey(data, txPowerField + '._value')) {
-        success = true;
-        ponSignal = {
-          rxpower: getFromNestedKey(data, rxPowerField + '._value'),
-          txpower: getFromNestedKey(data, txPowerField + '._value'),
-        };
-      } else if (checkForNestedKey(data, rxPowerFieldEpon + '._value') &&
-                 checkForNestedKey(data, txPowerFieldEpon + '._value')) {
-        success = true;
-        ponSignal = {
-          rxpower: getFromNestedKey(data, rxPowerFieldEpon + '._value'),
-          txpower: getFromNestedKey(data, txPowerFieldEpon + '._value'),
-        };
-      }
-      if (success) {
-        let deviceEdit = await DeviceModel.findById(mac);
-        if (!deviceEdit) return;
-        deviceEdit.last_contact = Date.now();
-        if (ponSignal.rxpower) {
-          ponSignal.rxpower = convertToDbm(deviceEdit.model, ponSignal.rxpower);
-        }
-        if (ponSignal.txpower) {
-          ponSignal.txpower = convertToDbm(deviceEdit.model, ponSignal.txpower);
-        }
-        ponSignal = appendPonSignal(
-          deviceEdit.pon_signal_measure,
-          ponSignal.rxpower,
-          ponSignal.txpower,
-        );
-        deviceEdit.pon_signal_measure = ponSignal;
-        await deviceEdit.save().catch((err) => {
-          console.log('Error saving pon signal: ' + err);
-        });
-      }
-      sio.anlixSendPonSignalNotification(mac, {ponsignalmeasure: ponSignal});
-      return ponSignal;
-    });
-  });
-  req.end();
-};
-
-// TODO: Move this function to external-genieacs?
-const fetchDevicesFromGenie = async function(acsID) {
-  let device;
-  try {
-    device = await DeviceModel.findOne({acs_id: acsID}).lean();
-  } catch (e) {
-    return;
-  }
-  if (!device || !device.use_tr069) {
-    return;
-  }
-  let mac = device._id;
-  let splitID = acsID.split('-');
-  let model = splitID.slice(1, splitID.length-1).join('-');
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  let hostsField = fields.devices.hosts;
-  let assocField = fields.devices.associated;
-  assocField = assocField.split('.').slice(0, -2).join('.');
-  let query = {_id: acsID};
-  let projection = hostsField + ',' + assocField;
-  let path = '/devices/?query='+JSON.stringify(query)+'&projection='+projection;
-  let options = {
-    method: 'GET',
-    hostname: 'localhost',
-    port: 7557,
-    path: encodeURI(path),
-  };
-  let req = http.request(options, (resp)=>{
-    resp.setEncoding('utf8');
-    let data = '';
-    resp.on('data', (chunk)=>data+=chunk);
-    resp.on('end', async () => {
-      if (data.length > 0) {
-        try {
-          data = JSON.parse(data)[0];
-        } catch (err) {
-          debug(err);
-          data = '';
-        }
-      }
-      let success = true;
-      let hostKeys = [];
-      let hostCountField = hostsField+'.HostNumberOfEntries._value';
-      // Make sure we have a host count and assodicated devices fields
-      if (checkForNestedKey(data, hostCountField) &&
-          checkForNestedKey(data, assocField)) {
-        getFromNestedKey(data, hostCountField);
-        // Host indexes might not respect order because of expired leases, so
-        // we just use whatever keys show up
-        let hostBaseField = fields.devices.hosts_template;
-        let hostKeysRaw = getFromNestedKey(data, hostBaseField);
-        if (hostKeysRaw) {
-          hostKeys = Object.keys(hostKeysRaw);
-        }
-        // Filter out meta fields from genieacs
-        hostKeys = hostKeys.filter((k)=>k[0] && k[0]!=='_');
-      } else {
-        success = false;
-      }
-      if (success) {
-        let iface2 = fields.wifi2.ssid.replace('.SSID', '');
-        let iface5 = fields.wifi5.ssid.replace('.SSID', '');
-        let devices = [];
-        hostKeys.forEach((i)=>{
-          let device = {};
-          // Collect device mac
-          let macKey = fields.devices.host_mac.replace('*', i);
-          device.mac = getFromNestedKey(data, macKey+'._value');
-          if (typeof device.mac === 'string') {
-            device.mac = device.mac.toUpperCase().replace(/-/g, ':');
-          } else {
-            // MAC is a mandatory string
-            return;
-          }
-          // Collect device hostname
-          let nameKey = fields.devices.host_name.replace('*', i);
-          device.name = getFromNestedKey(data, nameKey+'._value');
-          if (typeof device.name !== 'string' || device.name === '') {
-            // Needs a default name, use mac
-            device.name = device.mac;
-          }
-          // Collect device ip
-          let ipKey = fields.devices.host_ip.replace('*', i);
-          device.ip = getFromNestedKey(data, ipKey+'._value');
-          if (typeof device.ip !== 'string') {
-            // IP is mandatory
-            return;
-          }
-          // Collect layer 2 interface
-          let ifaceKey = fields.devices.host_layer2.replace('*', i);
-          let l2iface = getFromNestedKey(data, ifaceKey+'._value');
-          // DIR-842 leaves the dot in the layer2 interface name
-          if (l2iface === iface2 || l2iface === iface2 + '.') {
-            device.wifi = true;
-            device.wifi_freq = 2.4;
-          } else if (l2iface === iface5 || l2iface === iface5 + '.') {
-            device.wifi = true;
-            device.wifi_freq = 5;
-          }
-          // Push basic device information
-          devices.push(device);
-        });
-
-        if (fields.devices.host_rssi || fields.devices.host_snr) {
-          // Change iface identifiers to use only numerical identifier
-          iface2 = iface2.split('.');
-          iface5 = iface5.split('.');
-          iface2 = iface2[iface2.length-1];
-          iface5 = iface5[iface5.length-1];
-          // Filter wlan interfaces
-          let interfaces = Object.keys(getFromNestedKey(data, assocField));
-          interfaces = interfaces.filter((i)=>i[0]!='_');
-          if (fields.devices.associated_5) {
-            if (model == 'AC10') {
-              interfaces.push('2');
-            } else {
-              interfaces.push('5');
-            }
-          }
-          interfaces.forEach((iface) => {
-            // Get active indexes, filter metadata fields
-            assocField = fields.devices.associated.replace(
-              /WLANConfiguration\.[0-9*]+\./g,
-              'WLANConfiguration.' + iface + '.',
-            );
-            let assocIndexes = getFromNestedKey(data, assocField);
-            if (assocIndexes) {
-              assocIndexes = Object.keys(assocIndexes);
-            } else {
-              assocIndexes = [];
-            }
-            assocIndexes = assocIndexes.filter((i)=>i[0]!='_');
-            assocIndexes.forEach((index) => {
-              // Collect associated mac
-              let macKey = fields.devices.assoc_mac;
-              macKey = macKey.replace('*', iface).replace('*', index);
-              let macVal = getFromNestedKey(data, macKey+'._value');
-              if (typeof macVal === 'string') {
-                macVal = macVal.toUpperCase();
-              } else {
-                // MAC is mandatory
-                return;
-              }
-              let device = devices.find((d)=>d.mac.toUpperCase()===macVal);
-              if (!device) return;
-              // Mark device as a wifi device
-              device.wifi = true;
-              if (iface == iface2) {
-                device.wifi_freq = 2.4;
-              } else if (iface == iface5) {
-                device.wifi_freq = 5;
-              }
-              // Collect rssi, if available
-              if (fields.devices.host_rssi) {
-                let rssiKey = fields.devices.host_rssi;
-                rssiKey = rssiKey.replace('*', iface).replace('*', index);
-                device.rssi = getFromNestedKey(data, rssiKey+'._value');
-                // Casts to string if is a number so we can replace 'dBm'
-                if (typeof device.rssi === 'number') {
-                  device.rssi = device.rssi.toString();
-                }
-                device.rssi = device.rssi.replace('dBm', '');
-                // Cast back to number to avoid converting issues
-                device.rssi = parseInt(device.rssi);
-                if (isNaN(parseInt(device.rssi))) {
-                  debug(`device.rssi is NaN beware!!!`);
-                }
-              }
-              // Collect snr, if available
-              if (fields.devices.host_snr) {
-                let snrKey = fields.devices.host_snr;
-                snrKey = snrKey.replace('*', iface).replace('*', index);
-                device.snr = getFromNestedKey(data, snrKey+'._value');
-              } else if (fields.devices.host_rssi && device.rssi) {
-                device.snr = parseInt(device.rssi)+95;
-                if (isNaN(parseInt(device.rssi))) {
-                  debug(`device.rssi is NaN beware!!!`);
-                }
-              }
-              // Collect mode, if available
-              if (fields.devices.host_mode) {
-                let modeKey = fields.devices.host_mode;
-                modeKey = modeKey.replace('*', iface).replace('*', index);
-                let modeVal = getFromNestedKey(data, modeKey+'._value');
-                if (modeVal.includes('ac')) {
-                  device.wifi_mode = 'AC';
-                } else if (modeVal.includes('n')) {
-                  device.wifi_mode = 'N';
-                } else if (modeVal.includes('g')) {
-                  device.wifi_mode = 'G';
-                }
-              }
-              // Collect connection speed, if available
-              if (fields.devices.host_rate) {
-                let rateKey = fields.devices.host_rate;
-                rateKey = rateKey.replace('*', iface).replace('*', index);
-                device.rate = getFromNestedKey(data, rateKey+'._value');
-                device.rate = convertWifiRate(model, device.rate);
-              }
-              if (device.mac == device.name &&
-                fields.devices.alt_host_name) {
-                let nameKey = fields.devices.alt_host_name;
-                nameKey = nameKey.replace('*', iface).replace('*', index);
-                device.name = getFromNestedKey(data, nameKey+'._value');
-              }
-            });
-          });
-        }
-        await saveDeviceData(mac, devices).catch(debug);
-      }
-      sio.anlixSendOnlineDevNotifications(mac, null);
-    });
-  });
-  req.end();
+  return await acsDiagnosticsHandler.fireSpeedDiagnose(mac);
 };
 
 acsDeviceInfoController.requestLogs = function(device) {
@@ -2369,7 +1121,7 @@ acsDeviceInfoController.requestLogs = function(device) {
     name: 'getParameterValues',
     parameterNames: [logField],
   };
-  TasksAPI.addTask(acsID, task, fetchLogFromGenie);
+  TasksAPI.addTask(acsID, task, acsDeviceLogsHandler.fetchLogFromGenie);
 };
 
 acsDeviceInfoController.requestWanBytes = function(device) {
@@ -2386,7 +1138,7 @@ acsDeviceInfoController.requestWanBytes = function(device) {
       sentField,
     ],
   };
-  TasksAPI.addTask(acsID, task, fetchWanBytesFromGenie);
+  TasksAPI.addTask(acsID, task, acsMeasuresHandler.fetchWanBytesFromGenie);
 };
 
 acsDeviceInfoController.requestUpStatus = function(device) {
@@ -2417,7 +1169,7 @@ acsDeviceInfoController.requestUpStatus = function(device) {
       task.parameterNames.push(fields.wan.pon_txpower_epon);
     }
   }
-  TasksAPI.addTask(acsID, task, fetchUpStatusFromGenie);
+  TasksAPI.addTask(acsID, task, acsMeasuresHandler.fetchUpStatusFromGenie);
 };
 
 acsDeviceInfoController.requestConnectedDevices = function(device) {
@@ -2455,138 +1207,7 @@ const getSsidPrefixCheck = async function(device) {
 acsDeviceInfoController.getMeshBSSIDFromGenie = async function(
   device, meshMode,
 ) {
-  let acsID = device.acs_id;
-  // We have to check if the virtual AP object has been created already
-  const fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  const bssidField2 = fields.mesh2.bssid;
-  const bssidField5 = fields.mesh5.bssid;
-  const getObjTask = {
-    name: 'getParameterValues',
-    parameterNames: [],
-  };
-  if (meshMode === 2 || meshMode === 4) {
-    getObjTask.parameterNames.push(bssidField2);
-  }
-  if (meshMode === 3 || meshMode === 4) {
-    getObjTask.parameterNames.push(bssidField5);
-  }
-  let bssidsStatus;
-  try {
-    let ret = await TasksAPI.addTask(acsID, getObjTask);
-    if (!ret || !ret.success || !ret.executed) {
-      throw new Error('task error');
-    }
-    bssidsStatus = await fetchMeshBSSID(device, meshMode);
-    if (!bssidsStatus.success) {
-      throw new Error('invalid data');
-    }
-  } catch (e) {
-    const msg = `[!] -> ${e.message} in ${acsID}`;
-    console.log(msg);
-    return {success: false, msg: msg};
-  }
-  return {
-    success: true,
-    bssid_mesh2: bssidsStatus.mesh2.toUpperCase(),
-    bssid_mesh5: bssidsStatus.mesh5.toUpperCase(),
-  };
-};
-
-acsDeviceInfoController.createVirtualAPObjects = async function(device) {
-  let acsID = device.acs_id;
-  // We have to check if the virtual AP object has been created already
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  const meshField = fields.mesh2.ssid.replace('.SSID', '');
-  const meshField5 = fields.mesh5.ssid.replace('.SSID', '');
-  const getObjTask = {
-    name: 'getParameterValues',
-    parameterNames: [meshField, meshField5],
-  };
-  let meshObjsStatus;
-  try {
-    let ret = await TasksAPI.addTask(acsID, getObjTask);
-    if (!ret || !ret.success || !ret.executed) {
-      throw new Error('task error');
-    }
-    if (ret.finished) {
-      meshObjsStatus = await checkMeshObjsCreated(device);
-      if (!meshObjsStatus.success) {
-        throw new Error('invalid data');
-      }
-    }
-  } catch (e) {
-    const msg = `[!] -> ${e.message} in ${acsID}`;
-    console.log(msg);
-    return {success: false, msg: msg};
-  }
-  let deleteMesh5VAP = false;
-  let createMesh2VAP = false;
-  let createMesh5VAP = false;
-  /*
-    If the 2.4GHz virtual AP object hasn't been created we must create it.
-    Since the objects are created in order we must delete the 5GHz virtual AP
-    object if it exists and then recreate it. We never delete the 2.4GHz VAP
-    object, only the 5GHz one in specific cases
-  */
-  if (!meshObjsStatus.mesh2) {
-    createMesh2VAP = true;
-    createMesh5VAP = true;
-    if (meshObjsStatus.mesh5) {
-      deleteMesh5VAP = true;
-    }
-  } else if (!meshObjsStatus.mesh5) {
-    // 2.4GHz virtual AP object is created. Here we treat only the 5GHz case.
-    createMesh5VAP = true;
-  }
-  if (deleteMesh5VAP) {
-    let delObjTask = {name: 'deleteObject', objectName: meshField5};
-    try {
-      let ret = await TasksAPI.addTask(acsID, delObjTask);
-      if (!ret || !ret.success || !ret.executed) {
-        throw new Error('delObject task error');
-      }
-    } catch (e) {
-      const msg = `[!] -> ${e.message} in ${acsID}`;
-      console.log(msg);
-      return {sucess: false, msg: msg};
-    }
-  }
-  // Virtual APs objects haven't been created yet - do so now
-  if (createMesh2VAP || createMesh5VAP) {
-    let addObjTask = {
-      name: 'addObject',
-      // Removes index of the WLANConfiguration field name.
-      // Will work only if 2.4GHz VAP WLANConfiguration index is lower than 10
-      objectName: meshField.slice(0, -2),
-    };
-    let numObjsToCreate = createMesh2VAP + createMesh5VAP; // cast bool to int
-    for (let i = 0; i < numObjsToCreate; i++) {
-      try {
-        let ret = await TasksAPI.addTask(acsID, addObjTask, true,
-          10000, [5000, 10000]);
-        if (!ret || !ret.success || !ret.executed) {
-          throw new Error('task error');
-        }
-      } catch (e) {
-        const msg = `[!] -> ${e.message} in ${acsID}`;
-        console.log(msg);
-        return {success: false, msg: msg};
-      }
-      // A getParameterValues call forces the whole object to be created
-      try {
-        let ret = await TasksAPI.addTask(acsID, getObjTask);
-        if (!ret || !ret.success || !ret.executed) {
-          throw new Error('task error');
-        }
-      } catch (e) {
-        const msg = `[!] -> ${e.message} in ${acsID}`;
-        console.log(msg);
-        return {success: false, msg: msg};
-      }
-    }
-  }
-  // We must populate the newly created fields, signal that to next function
-  return {success: true, populate: (createMesh2VAP || createMesh5VAP)};
+  return await acsMeshDeviceHandler.getMeshBSSIDFromGenie(device, meshMode);
 };
 
 acsDeviceInfoController.updateInfo = async function(
@@ -2748,768 +1369,16 @@ acsDeviceInfoController.updateInfo = async function(
   }
 };
 
-acsDeviceInfoController.changePortForwardRules = async function(device,
-                                                                rulesDiffLength,
+acsDeviceInfoController.changePortForwardRules = async function(
+  device, rulesDiffLength,
 ) {
-  // Make sure we only work with TR-069 devices with a valid ID
-  if (!device || !device.use_tr069 || !device.acs_id) return;
-  let i;
-  let ret;
-  // let mac = device._id;
-  let acsID = device.acs_id;
-  let splitID = acsID.split('-');
-  let model = splitID.slice(1, splitID.length-1).join('-');
-  // redirect to config file binding instead of setParametervalues
-  if (model == 'GONUAC001' || model == 'xPON' ||
-      model == 'HG9') {
-    configFileEditing(device, ['port-forward']);
-    return;
-  }
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  let changeEntriesSizeTask = {name: 'addObject', objectName: ''};
-  let updateTasks = {name: 'setParameterValues', parameterValues: []};
-  let portMappingTemplate = '';
-  if (device.connection_type === 'pppoe') {
-    portMappingTemplate = fields.port_mapping_ppp;
-  } else {
-    portMappingTemplate = fields.port_mapping_dhcp;
-  }
-  // check if already exists add, delete, set sent tasks
-  // getting older tasks for this device id.
-  let query = {device: acsID}; // selecting all tasks for a given device id.
-  let tasks;
-  try {
-    tasks = await TasksAPI.getFromCollection('tasks', query);
-  } catch (e) {
-    console.log('[!] -> '+e.message+' in '+acsID);
-  }
-  if (!Array.isArray(tasks)) return;
-  // if find some task with name addObject or deleteObject
-  let hasAlreadySentTasks = tasks.some((t) => {
-    return t.name === 'addObject' ||
-    t.name === 'deleteObject';
-  });
-  // drop this call of changePortForwardRules
-  if (hasAlreadySentTasks) {
-    console.log('[#] -> DC in '+acsID);
-    return;
-  }
-  // change array size via addObject or deleteObject
-  if (rulesDiffLength < 0) {
-    rulesDiffLength = -rulesDiffLength;
-    changeEntriesSizeTask.name = 'deleteObject';
-    for (i = (device.port_mapping.length + rulesDiffLength);
-        i > device.port_mapping.length;
-        i--) {
-      changeEntriesSizeTask.objectName = portMappingTemplate + '.' + i;
-      try {
-        ret = await TasksAPI.addTask(acsID, changeEntriesSizeTask);
-        if (!ret || !ret.success || !ret.executed) {
-          return;
-        }
-      } catch (e) {
-        console.log('[!] -> '+e.message+' in '+acsID);
-      }
-    }
-    console.log('[#] -> D('+rulesDiffLength+') in '+acsID);
-  } else if (rulesDiffLength > 0) {
-    changeEntriesSizeTask.objectName = portMappingTemplate;
-    for (i = 0; i < rulesDiffLength; i++) {
-      try {
-        ret = await TasksAPI.addTask(acsID, changeEntriesSizeTask);
-        if (!ret || !ret.success || !ret.executed) {
-          return;
-        }
-      } catch (e) {
-        console.log('[!] -> '+e.message+' in '+acsID);
-      }
-    }
-    console.log('[#] -> A('+rulesDiffLength+') in '+acsID);
-  }
-  // set entries values for respective array in the device
-  for (i = 0; i < device.port_mapping.length; i++) {
-    const iterateTemplate = portMappingTemplate + '.' + (i+1) + '.';
-    Object.entries(fields.port_mapping_fields).forEach((v) => {
-      updateTasks.parameterValues.push([
-        iterateTemplate+v[1][0],
-        device.port_mapping[i][v[1][1]], v[1][2]]);
-    });
-    Object.entries(fields.port_mapping_values).forEach((v) => {
-      if (v[0] == 'description') {
-        v[1][1] = 'Anlix_PortForwarding_'+(i+1).toString();
-      }
-      updateTasks.parameterValues.push([
-        iterateTemplate+v[1][0], v[1][1], v[1][2]]);
-    });
-  }
-  // just send tasks if there are port mappings to fill/set
-  if (updateTasks.parameterValues.length > 0) {
-    console.log('[#] -> U in '+acsID);
-    TasksAPI.addTask(acsID, updateTasks);
-  }
+  return await acsPortForwardHandler.changePortForwardRules(
+    device, rulesDiffLength,
+  );
 };
 
 acsDeviceInfoController.changeAcRules = async function(device) {
-  // ===== Check device =====
-  // Make sure we only work with TR-069 devices with a valid ID
-  if (!device || !device.use_tr069 || !device.acs_id) return;
-  let acsID = device.acs_id;
-  let serial = device.serial_tr069;
-  // Make sure that this device is abled to do access control
-  let permissions = DeviceVersion.findByVersion(
-    device.version, device.wifi_is_5ghz_capable, device.model,
-  );
-  if (!permissions || !permissions.grantBlockDevices) return;
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  if (!fields) return;
-  // ===== Get blockedDevices and AC rules trees =====
-  // Creates the structures related to WLAN subtrees
-  // Make sure there are no more than 31 devices to block - limit of 64 rules
-  let blockedDevicesCount = Object.values(device.lan_devices).reduce((v, d)=>{
-    return v + ((d.is_blocked) ? 1 : 0);
-  }, 0);
-  if (blockedDevicesCount >= 32) {
-    console.log('Number of rules has reached its limit for device ' + acsID);
-    return {
-      success: false,
-      error_code: 'acRuleLimits',
-      message: t('acRuleLimits', {errorline: __line})
-    };
-  }
-  let acSubtreeRoots = {'wifi2': fields.access_control.wifi2};
-  if (device.wifi_is_5ghz_capable) {
-    acSubtreeRoots['wifi5'] = fields.access_control.wifi5;
-  }
-  // ===== Update device tree =====
-  let task = {
-    name: 'getParameterValues',
-    parameterNames: Object.values(acSubtreeRoots),
-  };
-  result = await TasksAPI.addTask(acsID, task, compareNewACRulesWithTree);
-  if (!result || !result.success) {
-    console.log('Error: failed to retrieve Access Control Rules at '+serial);
-    return {
-      'success': false, 'error_code': 'acRuleGetError',
-      'message': t('acRuleGetError', {errorline: __line}),
-    };
-  }
-  return {'success': true};
-};
-
-const compareNewACRulesWithTree = async function(acsID) {
-  let device;
-  try {
-    device = await DeviceModel.findOne({acs_id: acsID}).lean();
-  } catch (e) {
-    return;
-  }
-  if (!device || !device.use_tr069) {
-    return;
-  }
-
-  let maxId = 0;
-  let acRulesResult = {};
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  let blockedDevices = Object.values(device.lan_devices).filter(
-    (d)=>d.is_blocked,
-  );
-
-  let supportedWlans = ['wifi2'];
-  let acSubtreeRoots = {'wifi2': fields.access_control.wifi2};
-  let wlanAcRulesTrees = {'wifi2': []};
-  if (device.wifi_is_5ghz_capable) {
-    supportedWlans.push('wifi5');
-    acSubtreeRoots['wifi5'] = fields.access_control.wifi5;
-    wlanAcRulesTrees['wifi5'] = [];
-  }
-
-  try {
-    acRulesResult = await getAcRuleTrees(
-      acsID, acSubtreeRoots, supportedWlans, wlanAcRulesTrees, maxId,
-    );
-    if (!('success' in acRulesResult) || !acRulesResult.success) {
-      console.log('Error getting AC rules for device ' + acsID);
-      return;
-    }
-  } catch (e) {
-    console.log('Exception getting AC rules for device ' + acsID);
-    return;
-  }
-
-  let rulesToEdit = {'wifi2': {}};
-  let rulesToDelete = {'wifi2': {}};
-  if (device.wifi_is_5ghz_capable) {
-    rulesToEdit['wifi5'] = {};
-    rulesToDelete['wifi5'] = {};
-  }
-  let ids = acRulesResult['rule_ids'];
-  wlanAcRulesTrees = acRulesResult['rule_trees'];
-  maxId = acRulesResult['max_id'];
-  // If there is an imbalance in the tree or if the tree has a gap, it will
-  // delete all the rules in the tree and force the algorithm to re-populate
-  // all the rules in the tree.
-  // If the algorithm got this far and it has a rule with the id equal to
-  // the limit of 64, then there is a problem, because there are not 64
-  // blocked devices. It will also force the algorithm to repopulate the
-  // rule tree.
-  let condA = !device.wifi_is_5ghz_capable && ids['wifi2'].length != maxId;
-  let condBA = (ids['wifi2'].length + ids['wifi5'].length) != maxId;
-  let condBB = ids['wifi2'].length != ids['wifi5'].length;
-  let condB = condBA || condBB;
-  let condC = device.wifi_is_5ghz_capable;
-  if (condA || (condB && condC) || maxId >= 64) {
-    for (let wlanType of supportedWlans) {
-      let deleteRulesResult = await deleteAcRules(
-        device, wlanAcRulesTrees[wlanType],
-      );
-      if (!deleteRulesResult) {
-        console.log('Error deleting AC rules for redo device ' + acsID);
-        return;
-      }
-    }
-    maxId = 0;
-    wlanAcRulesTrees = {'wifi2': []};
-    if (permissions.grantWifi5ghz) {
-      wlanAcRulesTrees['wifi5'] = [];
-    }
-  }
-  let diff = blockedDevices.length - (maxId/2);
-  if (!device.wifi_is_5ghz_capable) {
-    diff = blockedDevices.length - maxId;
-  }
-  // If the difference is positive then you need to add rules to the tree.
-  if (diff > 0) {
-    rulesToEdit = wlanAcRulesTrees;
-    let addedRules;
-    try {
-      let addRulesResult =
-        await addAcRules(
-          device, acSubtreeRoots, diff, maxId, device.wifi_is_5ghz_capable,
-        );
-      if ('success' in addRulesResult && addRulesResult['success']) {
-        addedRules = addRulesResult['added_rules'];
-      }
-    } catch (e) {
-      console.log('Error ('+e+') at AC rules add');
-      return;
-    }
-    // Get the rules that need to be edited (old + new).
-    for (let wlanType of supportedWlans) {
-      if (rulesToEdit[wlanType].length > 0) {
-        rulesToEdit[wlanType] =
-          rulesToEdit[wlanType].concat(addedRules[wlanType]);
-      } else {
-        rulesToEdit[wlanType] = addedRules[wlanType];
-      }
-    }
-  // If the difference is negative, then we need to remove rules from the
-  // tree. The algorithm will reserve the lowest id rules, which are at the
-  // beginning of the rules array, and will delete as many as necessary
-  // until it has the right number of rules.
-  } else if (diff < 0) {
-    for (let wlanType of supportedWlans) {
-      let deleteSuccess = true;
-      let wlanTreeRules = wlanAcRulesTrees[wlanType];
-      // This sort solves the TR-069 sorting problem (eg 1, 11, 2, 3, ..., 10)
-      wlanTreeRules = wlanTreeRules.sort((a, b) => {
-        let aArr = a.split('.');
-        let aId = parseInt(aArr[aArr.length - 1], 10);
-        if (isNaN(aId)) {
-          debug(`wlanTreeRules aId is not an number`);
-        }
-        let bArr = b.split('.');
-        let bId = parseInt(bArr[bArr.length - 1], 10);
-        if (isNaN(bId)) {
-          debug(`wlanTreeRules bId is not an number`);
-        }
-        return aId - bId;
-      });
-      let wlanNumOfRules = wlanTreeRules.length;
-      rulesToEdit[wlanType] =
-        wlanTreeRules.slice(0, wlanNumOfRules+diff);
-      rulesToDelete =
-        wlanTreeRules.slice(wlanNumOfRules+diff, wlanNumOfRules);
-      try {
-        let deleteRulesResult =
-          await deleteAcRules(device, rulesToDelete);
-        if (!deleteRulesResult) {
-          deleteSuccess = false;
-        }
-      } catch (e) {
-        console.log('Error ('+e+') at AC rules delete');
-        deleteSuccess = false;
-      }
-      if (!deleteSuccess) {
-        return;
-      }
-    }
-  // If the difference is zero then we just need to edit the existing rules.
-  } else {
-    rulesToEdit = wlanAcRulesTrees;
-  }
-  // Checks if the rules in the TR-069 tree are in accordance with the
-  // number of blocked devices. For devices without 5Ghz, check step needs
-  // to be different
-  let allOkWithWifi2 = (rulesToEdit['wifi2'] && (
-    rulesToEdit['wifi2'].length == blockedDevices.length));
-  let allOkWithWifi5 = (permissions.grantWifi5ghz && (
-    rulesToEdit['wifi5'] && (
-    rulesToEdit['wifi5'].length == blockedDevices.length)));
-  // At this point the number of rules that exist in each WLAN tree and the
-  // number of blocked devices must be equal. Otherwise, it returns an error
-  if (allOkWithWifi2) {
-    if (!permissions.grantWifi5ghz || allOkWithWifi5) {
-      try {
-        await updateAcRules(
-          device, supportedWlans, acSubtreeRoots, rulesToEdit, blockedDevices
-        );
-        console.log('Updated Access Control tree successfully');
-      } catch (e) {
-        console.log('Error ('+e+') at AC rules update');
-      }
-    }
-  } else {
-    console.log('Error at AC rules update. Number of rules doesn\'t match');
-  }
-};
-
-const getAcRuleTrees = async function(
-  acsID, acSubtreeRoots, supportedWlans, wlanAcRulesTrees, maxId,
-) {
-  let wlanAcRulesIds = {};
-  // ===== Get device AC tree =====
-  let query = {_id: acsID};
-  let path = '/devices/?query='+JSON.stringify(query)+
-    '&projection='+Object.values(acSubtreeRoots).join(',');
-  let options = {
-    method: 'GET',
-    hostname: 'localhost',
-    port: 7557,
-    path: encodeURI(path),
-  };
-  // Projection Promise
-  return new Promise(function(resolve, reject) {
-    let req = http.request(options, (resp)=>{
-      resp.setEncoding('utf8');
-      let data = '';
-      resp.on('error', (error) => {
-        console.log('Error ('+ error +') removing an AC rule at '+acsID);
-        reject({
-          'success': false, 'error_code': 'acRuleGetError',
-          'message': t('acRuleGetError', {errorline: __line}),
-        });
-      });
-      resp.on('data', (chunk) => data+=chunk);
-      resp.on('end', async () => {
-        if (data.length > 0) {
-          try {
-            // Get data
-            data = JSON.parse(data)[0];
-            for (let wlanType of supportedWlans) {
-              let wlanTreeRoot = acSubtreeRoots[wlanType];
-              let wlanTreeRules = [];
-              let wlanTreeRuleIds = [];
-              // Populates the WLAN subtree rule structure
-              if (checkForNestedKey(data, wlanTreeRoot)) {
-                let wlanSubtree = getFromNestedKey(data, wlanTreeRoot);
-                Object.entries(wlanSubtree).forEach((acRule) => {
-                  let acRuleId = acRule[0];
-                  if (acRuleId && !acRuleId.startsWith('_')) {
-                    maxId = Math.max(Number(acRuleId), maxId);
-                    wlanTreeRuleIds.push(acRuleId);
-                    wlanTreeRules.push(wlanTreeRoot+'.'+acRuleId);
-                  }
-                });
-              }
-              wlanAcRulesIds[wlanType] = wlanTreeRuleIds;
-              wlanAcRulesTrees[wlanType] = wlanTreeRules;
-            }
-            // If success, Resolve
-            console.log('Successfully obtained AC rule trees');
-            return resolve({
-              'success': true,
-              'max_id': maxId,
-              'rule_ids': wlanAcRulesIds,
-              'rule_trees': wlanAcRulesTrees,
-            });
-          } catch (e) {
-            // If error, Reject
-            console.log(
-              'Error ('+e+') retrieving Access Control Rules at '+acsID);
-            return reject({
-              'success': false, 'error_code': 'acRuleGetError',
-              'message': t('acRuleGetError', {errorline: __line}),
-            });
-          }
-        }
-      });
-    });
-    req.end();
-  });
-};
-
-// Adds a new rule in the rules tree of each WLAN and assumes an id for this
-// rule from the id of the last rule.
-// This makes it possible for us to save a get and a projection.
-const addAcRules = async function(
-  device, acSubtreeRoots, numberOfRules, maxId, grantWifi5ghz
-) {
-  let currentMaxId = parseInt(maxId, 10);
-  if (isNaN(currentMaxId)) {
-    debug('maxId is not an number');
-  }
-  let acRulesAdded = {'wifi2': []};
-  if (grantWifi5ghz) {
-    acRulesAdded['wifi5'] = [];
-  }
-  for (let i = 0; i < numberOfRules; i++) {
-    // Adding at wifi2 WLAN
-    currentMaxId = currentMaxId + 1;
-    let result = await TasksAPI.addOrDeleteObject(
-      device.acs_id, acSubtreeRoots['wifi2'], 'addObject');
-    if (result) {
-      acRulesAdded['wifi2'].push(
-        acSubtreeRoots['wifi2']+'.'+currentMaxId.toString(),
-      );
-    } else {
-      console.log('addAcRules error');
-      return {'success': false};
-    }
-    if (grantWifi5ghz) {
-      // Adding at wifi5 WLAN
-      currentMaxId = currentMaxId + 1;
-      result = await TasksAPI.addOrDeleteObject(
-        device.acs_id, acSubtreeRoots['wifi5'], 'addObject');
-      if (result) {
-        acRulesAdded['wifi5'].push(
-          acSubtreeRoots['wifi5']+'.'+currentMaxId.toString(),
-        );
-      // If we were not successful adding a rule to the subtree of wifi5 WLAN,
-      // then we need to rollback the added rule to wifi2 WLAN.
-      } else {
-        await deleteAcRules(device, acRulesAdded['wifi2'].slice(-1));
-        console.log('addAcRules error');
-        return {'success': false};
-      }
-    }
-  }
-  return {'success': true, 'added_rules': acRulesAdded};
-};
-
-const deleteAcRules = async function(device, rulesToDelete) {
-  try {
-    for (let acRule of rulesToDelete) {
-      let result = await TasksAPI.addOrDeleteObject(
-        device.acs_id, acRule, 'deleteObject'
-      );
-      if (!result) {
-        console.log('deleteAcRules error');
-        return false;
-      }
-    }
-  } catch (e) {
-    console.log('Exception deleting AC rules for device ' + device.acs_id);
-    return false;
-  }
-  return true;
-};
-
-const updateAcRules = async function(
-  device, supportedWlans, acSubtreeRoots, rulesToEdit, blockedDevices,
-) {
-  let acsID = device.acs_id;
-  let serial = device.serial_tr069;
-  let parameterValues = [];
-  try {
-    for (let wlanType of supportedWlans) {
-      if (rulesToEdit[wlanType].length !== blockedDevices.length) {
-        return;
-      }
-      let acSubtreeRoot = acSubtreeRoots[wlanType];
-      parameterValues.push([acSubtreeRoot+'Mode', 'Ban', 'xsd:string']);
-      for (let i = 0; i < blockedDevices.length; i++) {
-        let ruleArr = (rulesToEdit[wlanType][i]).split('.');
-        let ruleId = ruleArr[ruleArr.length - 1];
-        // While we need to add rules to block devices, we fill the name and MAC
-        // fields with the necessary values.
-        let ruleMac = blockedDevices[i]['mac'];
-        // Generating a random string combined with last 4 digits of devices mac
-        let ruleName = ruleId + '-' + ruleMac.replace(/:/g, '').slice(-4);
-        // Just one more check step. This model only accepts 10 caracteres in
-        // the name field.
-        ruleName = (ruleName.length > 10) ? ruleName.slice(-10) : ruleName;
-        parameterValues.push(
-          [rulesToEdit[wlanType][i]+'.'+'Name', ruleName, 'xsd:string'],
-          [rulesToEdit[wlanType][i]+'.'+'MACAddress', ruleMac, 'xsd:string'],
-        );
-      }
-    }
-  } catch (e) {
-    return;
-  }
-  if (parameterValues.length > 0) {
-    try {
-      let task = {
-        name: 'setParameterValues',
-        parameterValues: parameterValues,
-      };
-      let result = await TasksAPI.addTask(acsID, task);
-      if (!result || !result.success) {
-        console.log('Error in editing Access Control rules task at '+serial);
-      }
-    } catch (e) {
-      console.log('Error: failed to edit Access Control rules at '+serial);
-    }
-  }
-};
-
-const configFileEditing = async function(device, target) {
-  let acsID = device.acs_id;
-  let serial = device.serial_tr069;
-  // get xml config file to genieacs
-  let configField = 'InternetGatewayDevice.DeviceConfig.ConfigFile';
-  let task = {
-    name: 'getParameterValues',
-    parameterNames: [configField],
-  };
-  let cback = (acsID)=>fetchAndEditConfigFile(acsID, target);
-  let result = await TasksAPI.addTask(acsID, task, cback);
-  if (!result || !result.success) {
-    console.log('Error: failed to retrieve ConfigFile at '+serial);
-    return;
-  }
-};
-
-const fetchAndEditConfigFile = async function(acsID, target) {
-  let device;
-  try {
-    device = await DeviceModel.findOne({acs_id: acsID}).lean();
-  } catch (e) {
-    return;
-  }
-  if (!device || !device.use_tr069) {
-    return;
-  }
-  let serial = device.serial_tr069;
-  let configField = 'InternetGatewayDevice.DeviceConfig.ConfigFile';
-
-  // get xml config file from genieacs
-  let query = {_id: acsID};
-  let path = '/devices/?query='+JSON.stringify(query)+
-    '&projection='+configField;
-  let options = {
-    method: 'GET',
-    hostname: 'localhost',
-    port: 7557,
-    path: encodeURI(path),
-  };
-  let req = http.request(options, (resp)=>{
-    resp.setEncoding('utf8');
-    let rawConfigFile = '';
-    resp.on('data', (chunk)=>rawConfigFile+=chunk);
-    resp.on('end', async () => {
-      if (rawConfigFile.length > 0) {
-        try {
-          rawConfigFile = JSON.parse(rawConfigFile)[0];
-        } catch (err) {
-          debug(err);
-          rawConfigFile = '';
-        }
-      }
-      if (checkForNestedKey(rawConfigFile, configField+'._value')) {
-        // modify xml config file
-        rawConfigFile = getFromNestedKey(rawConfigFile, configField+'._value');
-        let xmlConfigFile = acsHandlers
-          .digestXmlConfig(device, rawConfigFile, target);
-        if (xmlConfigFile != '') {
-          // set xml config file to genieacs
-          let task = {
-            name: 'setParameterValues',
-            parameterValues: [[configField, xmlConfigFile, 'xsd:string']],
-          };
-          let result = await TasksAPI.addTask(acsID, task);
-          if (!result || !result.success) {
-            console.log('Error: failed to write ConfigFile at '+serial);
-            return;
-          }
-        } else {
-          console.log('Error: failed xml validation at '+serial);
-        }
-      } else {
-        console.log('Error: no config file retrieved at '+serial);
-      }
-    });
-  });
-  req.end();
-};
-
-acsDeviceInfoController.checkPortForwardRules = async function(device) {
-  if (!device || !device.use_tr069 || !device.acs_id) return;
-  let acsID = device.acs_id;
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  let task = {
-    name: 'getParameterValues',
-    parameterNames: [],
-  };
-  let portMappingTemplate = '';
-  if (device.connection_type === 'pppoe') {
-    portMappingTemplate = fields.port_mapping_ppp;
-  } else {
-    portMappingTemplate = fields.port_mapping_dhcp;
-  }
-  task.parameterNames.push(portMappingTemplate);
-  let result = await TasksAPI.addTask(acsID, task, fetchAndComparePortForward);
-  if (!result || !result.success) {
-    console.log('Error getting port forward fields on device ' + acsID);
-  }
-};
-
-const fetchAndComparePortForward = async function(acsID) {
-  let device;
-  try {
-    device = await DeviceModel.findOne({acs_id: acsID}).lean();
-  } catch (e) {
-    return;
-  }
-  if (!device || !device.use_tr069) {
-    return;
-  }
-  let fields = DevicesAPI.getModelFieldsFromDevice(device).fields;
-  let portMappingTemplate = '';
-  if (device.connection_type === 'pppoe') {
-    portMappingTemplate = fields.port_mapping_ppp;
-  } else {
-    portMappingTemplate = fields.port_mapping_dhcp;
-  }
-
-  let query = {_id: acsID};
-  let projection1 = portMappingTemplate.replace('*', '1').replace('*', '1');
-  let projection2 = portMappingTemplate.replace('*', '1').replace('*', '2');
-  let path = '/devices/?query=' + JSON.stringify(query) + '&projection=' +
-             projection1 + ',' + projection2;
-  let options = {
-    method: 'GET',
-    hostname: 'localhost',
-    port: 7557,
-    path: encodeURI(path),
-  };
-  let req = http.request(options, (resp) => {
-    resp.setEncoding('utf8');
-    let data = '';
-    let i;
-    resp.on('data', (chunk)=>data+=chunk);
-    resp.on('end', async ()=>{
-      if (data.length > 0) {
-        data = JSON.parse(data)[0];
-      }
-      let isDiff = false;
-      let template = '';
-      if (checkForNestedKey(data, projection1)) {
-        template = projection1;
-      } else if (checkForNestedKey(data, projection2)) {
-        template = projection2;
-      }
-      if (template != '') {
-        for (i = 0; i < device.port_mapping.length; i++) {
-          let iterateTemplate = template+'.'+(i+1)+'.';
-          let portMapEnablePath = iterateTemplate +
-                                  fields.port_mapping_values.enable[0];
-          if (checkForNestedKey(data, portMapEnablePath)) {
-            if (getFromNestedKey(data, portMapEnablePath) != true) {
-              isDiff = true;
-              break;
-            }
-          }
-          let portMapLeasePath = iterateTemplate +
-                                  fields.port_mapping_values.lease[0];
-          if (checkForNestedKey(data, portMapLeasePath)) {
-            if (getFromNestedKey(data, portMapLeasePath) != 0) {
-              isDiff = true;
-              break;
-            }
-          }
-          let portMapProtocolPath = iterateTemplate +
-                                    fields.port_mapping_values.protocol[0];
-          if (checkForNestedKey(data, portMapProtocolPath)) {
-            if (getFromNestedKey(data,
-              portMapProtocolPath) != fields.port_mapping_values.protocol[1]
-            ) {
-              isDiff = true;
-              break;
-            }
-          }
-          let portMapDescriptionPath = iterateTemplate +
-                                    fields.port_mapping_values.description[0];
-          if (checkForNestedKey(data, portMapDescriptionPath)) {
-            if (
-              getFromNestedKey(data, portMapDescriptionPath) !=
-              fields.port_mapping_values.description[1]
-            ) {
-              isDiff = true;
-              break;
-            }
-          }
-          let portMapClientPath = iterateTemplate +
-                                  fields.port_mapping_fields.client[0];
-          if (checkForNestedKey(data, portMapClientPath)) {
-            if (getFromNestedKey(data,
-              portMapClientPath) != device.port_mapping[i].ip
-            ) {
-              isDiff = true;
-              break;
-            }
-          }
-          let portMapExtStart = iterateTemplate +
-            fields.port_mapping_fields.external_port_start[0];
-          if (checkForNestedKey(data, portMapExtStart)) {
-            if (getFromNestedKey(data, portMapExtStart) !=
-              device.port_mapping[i].external_port_start) {
-              isDiff = true;
-              break;
-            }
-          }
-          if (fields.port_mapping_fields.external_port_end) {
-            let portMapExtEnd = iterateTemplate +
-              fields.port_mapping_fields.external_port_end[0];
-            if (checkForNestedKey(data, portMapExtEnd)) {
-              if (getFromNestedKey(data, portMapExtEnd) !=
-                device.port_mapping[i].external_port_end) {
-                isDiff = true;
-                break;
-              }
-            }
-          }
-          let portMapIntStart = iterateTemplate +
-            fields.port_mapping_fields.internal_port_start[0];
-          if (checkForNestedKey(data, portMapIntStart)) {
-            if (getFromNestedKey(data, portMapIntStart) !=
-              device.port_mapping[i].internal_port_start) {
-              isDiff = true;
-              break;
-            }
-          }
-          if (fields.port_mapping_fields.internal_port_end) {
-            let portMapIntEnd = iterateTemplate +
-              fields.port_mapping_fields.internal_port_end[0];
-            if (checkForNestedKey(data, portMapIntEnd)) {
-              if (getFromNestedKey(data, portMapIntEnd) !=
-                device.port_mapping[i].internal_port_end) {
-                isDiff = true;
-                break;
-              }
-            }
-          }
-        }
-        if (isDiff) {
-          acsDeviceInfoController.changePortForwardRules(device, 0);
-        }
-      } else {
-        console.log('Wrong PortMapping in the device tree ' +
-                    'from genie. ACS ID is ' + acsID);
-      }
-    });
-  });
-  req.end();
+  return await acsAccessControlHandler.changeAcRules(device);
 };
 
 acsDeviceInfoController.pingOfflineDevices = async function() {
@@ -3614,61 +1483,21 @@ acsDeviceInfoController.reportOnuDevices = async function(app, devices=null) {
 };
 
 acsDeviceInfoController.addFirmwareInACS = async function(firmware) {
-  let binData;
-  try {
-    binData = await FirmwaresAPI.receiveFile(firmware.filename);
-  } catch (e) {
-    return false;
-  }
-  try {
-    await FirmwaresAPI.uploadToGenie(binData, firmware);
-  } catch (e) {
-    return false;
-  }
-  return true;
+  return await acsFirmwareHandler.addFirmwareInACS(firmware);
 };
 
 acsDeviceInfoController.delFirmwareInACS = async function(filename) {
-  await FirmwaresAPI.delFirmwareInGenie(filename);
+  return await acsFirmwareHandler.delFirmwareInACS(filename);
 };
 
 acsDeviceInfoController.upgradeFirmware = async function(device) {
-  let firmwares;
-  // verify existence in nbi through 7557/files/
-  firmwares = await FirmwaresAPI.getFirmwaresFromGenie();
-
-  let firmware = firmwares.find((f) => f.metadata.version == device.release);
-  // if not exists, then add
-  if (!firmware) {
-    firmware = await FirmwareModel.findOne({
-      model: device.model,
-      release: device.release,
-      cpe_type: 'tr069',
-    });
-    if (!firmware) {
-      return {success: false,
-        message: t('firmwareVersionNotFound', {errorline: __line})};
-    } else {
-      let response = await acsDeviceInfoController.addFirmwareInACS(firmware);
-      if (!response) {
-        return {success: false,
-          message: t('firmwareInsertError', {errorline: __line})};
-      }
-    }
-  }
-  // trigger 7557/devices/<acs_id>/tasks POST 'name': 'download'
-  let response = '';
-  try {
-    response = await FirmwaresAPI.sendUpgradeFirmware(firmware, device);
-  } catch (e) {
-    return {success: false, message: e.message};
-  }
-  return {success: true, message: response};
+  return await acsFirmwareHandler.upgradeFirmware(device);
 };
 
 // Should be called after validating mesh configuration
 acsDeviceInfoController.configTR069VirtualAP = async function(
-  device, targetMode) {
+  device, targetMode,
+) {
   const wifiRadioState = 1;
   const meshChannel = 7;
   const meshChannel5GHz = 40; // Value has better results on some routers
@@ -3688,8 +1517,7 @@ acsDeviceInfoController.configTR069VirtualAP = async function(
   */
   let createOk = {populate: false};
   if (!hasMeshVAPObject && targetMode > 0) {
-    createOk =
-      await acsDeviceInfoController.createVirtualAPObjects(device);
+    createOk = await acsMeshDeviceHandler.createVirtualAPObjects(device);
     if (!createOk.success) {
       return {success: false, msg: createOk.msg};
     }
